finito esercizio

+++ GestioneEventi/src/dao/IEventoDAO.java
@@ -5,7 +5,7 @@ import model.Evento;
 public interface IEventoDAO {
 	abstract void save(Evento e);
 
-	abstract Evento getById(int id);
+	abstract Evento getById(Long id);
 
-	abstract void delete(int id);
+	abstract void delete(Long id);
 }
